backend/requirements.txt
# VecQuery Intelligence — Python dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9        # required for FastAPI file uploads

# Database
sqlalchemy==2.0.30
psycopg[binary]==3.1.18        # PostgreSQL driver (pre-built wheels, Python 3.9 compatible)
pgvector==0.3.2                # pgvector SQLAlchemy type

# Document parsing
PyMuPDF==1.24.3                # PDF parsing (fitz)
python-docx==1.1.2             # DOCX parsing

# HTTP client (Ollama calls)
httpx==0.27.0

# Environment variables
python-dotenv==1.0.1
